+++ src/qf/qf_ps.c
@@ -203,7 +203,7 @@ void QF_publish_(QEvt const * const e, void const * const sender)
 * @sa QF_publish_(), QActive_unsubscribe(), and QActive_unsubscribeAll()
 */
 void QActive_subscribe(QActive const * const me, enum_t const sig) {
-    uint_fast8_t p = me->prio;
+    uint_fast8_t p = (uint_fast8_t)me->prio;
     QF_CRIT_STAT_
 
     Q_REQUIRE_ID(300, ((enum_t)Q_USER_SIG <= sig)
@@ -248,7 +248,7 @@ void QActive_subscribe(QActive const * const me, enum_t const sig) {
 * @sa QF_publish_(), QActive_subscribe(), and QActive_unsubscribeAll()
 */
 void QActive_unsubscribe(QActive const * const me, enum_t const sig) {
-    uint_fast8_t p = me->prio;
+    uint_fast8_t p = (uint_fast8_t)me->prio;
     QF_CRIT_STAT_
 
     /** @pre the singal and the prioriy must be in ragne, the AO must also
@@ -295,18 +295,20 @@ void QActive_unsubscribe(QActive const * const me, enum_t const sig) {
 * @sa QF_publish_(), QActive_subscribe(), and QActive_unsubscribe()
 */
 void QActive_unsubscribeAll(QActive const * const me) {
-    uint_fast8_t p = me->prio;
+    uint_fast8_t p = (uint_fast8_t)me->prio;
     enum_t sig;
 
     Q_REQUIRE_ID(500, ((uint_fast8_t)0 < p)
-                       && (p <= (uint_fast8_t)QF_MAX_ACTIVE)
+                        && (p <= (uint_fast8_t)QF_MAX_ACTIVE)
                        && (QF_active_[p] == me));
 
     for (sig = (enum_t)Q_USER_SIG; sig < QF_maxPubSignal_; ++sig) {
         QF_CRIT_STAT_
         QF_CRIT_ENTRY_();
-        if (QPSet_hasElement(&QF_PTR_AT_(QF_subscrList_, sig), p)) {
-            QPSet_remove(&QF_PTR_AT_(QF_subscrList_, sig), p);
+        if (QPSet_hasElement(&QF_PTR_AT_(QF_subscrList_, sig),
+                             (uint_fast8_t)p))
+        {
+            QPSet_remove(&QF_PTR_AT_(QF_subscrList_, sig), (uint_fast8_t)p);
 
             QS_BEGIN_NOCRIT_(QS_QF_ACTIVE_UNSUBSCRIBE,
                              QS_priv_.locFilter[AO_OBJ], me)
